+++ gulpfile.ts
@@ -3,21 +3,22 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import * as cp from 'child_process';
+import * as fse from 'fs-extra';
 import * as gulp from 'gulp';
 import * as path from 'path';
 import { gulp_installAzureAccount, gulp_webpack } from 'vscode-azureextensiondev';
 
 declare let exports: { [key: string]: unknown };
 
-function test(): cp.ChildProcess {
-    const env = process.env;
-
-    env.DEBUGTELEMETRY = '1';
-    env.CODE_TESTS_PATH = path.join(__dirname, 'dist/test');
-    return cp.spawn('node', ['./node_modules/vscode/bin/test'], { stdio: 'inherit', env });
+async function prepareForWebpack(): Promise<void> {
+    const mainJsPath: string = path.join(__dirname, 'main.js');
+    let contents: string = (await fse.readFile(mainJsPath)).toString();
+    contents = contents
+        .replace('out/src/extension', 'dist/extension.bundle')
+        .replace(', true /* ignoreBundle */', '');
+    await fse.writeFile(mainJsPath, contents);
 }
 
-exports['webpack-dev'] = () => gulp_webpack('development');
-exports['webpack-prod'] = () => gulp_webpack('production');
-exports.test = gulp.series(gulp_installAzureAccount, test);
+exports['webpack-dev'] = gulp.series(prepareForWebpack, () => gulp_webpack('development'));
+exports['webpack-prod'] = gulp.series(prepareForWebpack, () => gulp_webpack('production'));
+exports.preTest = gulp_installAzureAccount;