+++ src/commands/logstream/startStreamingLogs.ts
@@ -1,0 +1,23 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import * as appservice from 'vscode-azureappservice';
+import { IActionContext } from 'vscode-azureextensionui';
+import { SiteTreeItem } from "../../explorer/SiteTreeItem";
+import { WebAppTreeItem } from '../../explorer/WebAppTreeItem';
+import { ext } from '../../extensionVariables';
+import { enableFileLogging } from './enableFileLogging';
+
+export async function startStreamingLogs(context: IActionContext, node?: SiteTreeItem): Promise<void> {
+    if (!node) {
+        node = await ext.tree.showTreeItemPicker<WebAppTreeItem>(WebAppTreeItem.contextValue, context);
+    }
+
+    const verifyLoggingEnabled: () => Promise<void> = async (): Promise<void> => {
+        await enableFileLogging({ ...context, suppressAlreadyEnabledMessage: true }, node);
+    };
+
+    await appservice.startStreamingLogs(node.root.client, verifyLoggingEnabled, node.logStreamLabel);
+}