+++ src/explorer/SiteTreeItem.ts
@@ -6,12 +6,12 @@
 import * as WebSiteModels from 'azure-arm-website/lib/models';
 import * as fse from 'fs-extra';
 import * as path from 'path';
-import { MessageItem, window } from 'vscode';
+import { MessageItem } from 'vscode';
 import { AppSettingsTreeItem, AppSettingTreeItem, deleteSite, DeploymentsTreeItem, DeploymentTreeItem, FolderTreeItem, ISiteTreeRoot, LinuxRuntimes, LogFilesTreeItem, SiteClient, SiteFilesTreeItem } from 'vscode-azureappservice';
 import { AzExtTreeItem, AzureParentTreeItem, AzureTreeItem, DialogResponses, IActionContext } from 'vscode-azureextensionui';
 import * as constants from '../constants';
 import { ext } from '../extensionVariables';
-import { openUrl } from '../utils/openUrl';
+import { venvUtils } from '../utils/venvUtils';
 import { getWorkspaceSetting, updateWorkspaceSetting } from '../vsCodeConfig/settings';
 import { ConnectionsTreeItem } from './ConnectionsTreeItem';
 import { CosmosDBConnection } from './CosmosDBConnection';
@@ -73,10 +73,6 @@ export abstract class SiteTreeItem extends AzureParentTreeItem<ISiteTreeRoot> {
 
     public get id(): string {
         return this.root.client.id;
-    }
-
-    public async browse(): Promise<void> {
-        await openUrl(this.root.client.defaultHostUrl);
     }
 
     public async loadMoreChildrenImpl(_clearCache: boolean): Promise<AzureTreeItem<ISiteTreeRoot>[]> {
@@ -149,31 +145,24 @@ export abstract class SiteTreeItem extends AzureParentTreeItem<ISiteTreeRoot> {
     }
 
     public async promptScmDoBuildDeploy(fsPath: string, runtime: string, context: IActionContext): Promise<void> {
-        const dontShowAgainButton: MessageItem = { title: "No, and don't show again" };
-        const learnMoreButton: MessageItem = { title: 'Learn More' };
+        context.telemetry.properties.enableScmInput = "Canceled";
+
+        const learnMoreLink: string = 'https://aka.ms/Kwwkbd';
+
         const buildDuringDeploy: string = `Would you like to update your workspace configuration to run build commands on the target server? This should improve deployment performance.`;
-        let input: MessageItem | undefined = learnMoreButton;
-        while (input === learnMoreButton) {
-            input = await window.showInformationMessage(buildDuringDeploy, DialogResponses.yes, dontShowAgainButton, learnMoreButton);
-            if (input === learnMoreButton) {
-                await openUrl('https://aka.ms/Kwwkbd');
-            }
-        }
+        const input: MessageItem | undefined = await ext.ui.showWarningMessage(buildDuringDeploy, { modal: true, learnMoreLink }, DialogResponses.yes, DialogResponses.no);
+
         if (input === DialogResponses.yes) {
             await this.enableScmDoBuildDuringDeploy(fsPath, runtime);
             context.telemetry.properties.enableScmInput = "Yes";
         } else {
             await updateWorkspaceSetting(constants.configurationSettings.showBuildDuringDeployPrompt, false, fsPath);
-            context.telemetry.properties.enableScmInput = "No, and don't show again";
-        }
-
-        if (!context.telemetry.properties.enableScmInput) {
-            context.telemetry.properties.enableScmInput = "Canceled";
+            context.telemetry.properties.enableScmInput = "No";
         }
     }
 
     public async enableScmDoBuildDuringDeploy(fsPath: string, runtime: string): Promise<void> {
-        const zipIgnoreFolders: string[] = this.getIgnoredFoldersForDeployment(runtime);
+        const zipIgnoreFolders: string[] = await this.getIgnoredFoldersForDeployment(fsPath, runtime);
         let oldSettings: string[] | string | undefined = getWorkspaceSetting(constants.configurationSettings.zipIgnorePattern, fsPath);
         if (!oldSettings) {
             oldSettings = [];
@@ -212,21 +201,38 @@ export abstract class SiteTreeItem extends AzureParentTreeItem<ISiteTreeRoot> {
 
     }
 
-    private getIgnoredFoldersForDeployment(runtime: string): string[] {
+    private async getIgnoredFoldersForDeployment(fsPath: string, runtime: string): Promise<string[]> {
         let ignoredFolders: string[];
         switch (runtime) {
             case LinuxRuntimes.node:
                 ignoredFolders = ['node_modules{,/**}'];
                 break;
             case LinuxRuntimes.python:
+                let venvFsPaths: string[];
+                try {
+                    venvFsPaths = (await venvUtils.getExistingVenvs(fsPath)).map(venvPath => `${venvPath}{,/**}`);
+                } catch (error) {
+                    // if there was an error here, don't block-- just assume none could be detected
+                    venvFsPaths = [];
+                }
+
                 // list of Python ignorables are pulled from here https://github.com/github/gitignore/blob/master/Python.gitignore
                 // Byte-compiled / optimized / DLL files
                 ignoredFolders = ['__pycache__{,/**}', '*.py[cod]', '*$py.class',
                     // Distribution / packaging
                     '.Python{,/**}', 'build{,/**}', 'develop-eggs{,/**}', 'dist{,/**}', 'downloads{,/**}', 'eggs{,/**}', '.eggs{,/**}', 'lib{,/**}', 'lib64{,/**}', 'parts{,/**}', 'sdist{,/**}', 'var{,/**}',
-                    'wheels{,/**}', 'share/python-wheels{,/**}', '*.egg-info{,/**}', '.installed.cfg', '*.egg', 'MANIFEST',
-                    // Environments
-                    '.env{,/**}', '.venv{,/**}', 'env{,/**}', 'venv{,/**}', 'ENV{,/**}', 'env.bak{,/**}', 'venv.bak{,/**}'];
+                    'wheels{,/**}', 'share/python-wheels{,/**}', '*.egg-info{,/**}', '.installed.cfg', '*.egg', 'MANIFEST'];
+
+                // Virtual Environments
+                const defaultVenvPaths: string[] = ['.env{,/**}', '.venv{,/**}', 'env{,/**}', 'venv{,/**}', 'ENV{,/**}', 'env.bak{,/**}', 'venv.bak{,/**}'];
+                for (const venvPath of venvFsPaths) {
+                    // don't add duplicates
+                    if (!defaultVenvPaths.find(p => p === venvPath)) {
+                        defaultVenvPaths.push(venvPath);
+                    }
+                }
+
+                ignoredFolders = ignoredFolders.concat(defaultVenvPaths);
                 break;
             default:
                 ignoredFolders = [];