+++ test/nightly/global.resource.test.ts
@@ -5,7 +5,6 @@
 
 import { ResourceManagementClient } from 'azure-arm-resource';
 import { WebSiteManagementClient } from 'azure-arm-website';
-import { IHookCallbackContext } from 'mocha';
 import * as vscode from 'vscode';
 import { TestAzureAccount } from 'vscode-azureextensiondev';
 import { AzExtTreeDataProvider, AzureAccountTreeItem, createAzureClient, ext } from '../../extension.bundle';
@@ -15,37 +14,35 @@ export let testAccount: TestAzureAccount;
 export let webSiteClient: WebSiteManagementClient;
 export const resourceGroupsToDelete: string[] = [];
 
-suiteSetup(async function (this: IHookCallbackContext): Promise<void> {
-    if (!longRunningTestsEnabled) {
-        this.skip();
+suiteSetup(async function (this: Mocha.Context): Promise<void> {
+    if (longRunningTestsEnabled) {
+        this.timeout(2 * 60 * 1000);
+        testAccount = new TestAzureAccount(vscode);
+        await testAccount.signIn();
+        ext.azureAccountTreeItem = new AzureAccountTreeItem(testAccount);
+        ext.tree = new AzExtTreeDataProvider(ext.azureAccountTreeItem, 'appService.loadMore');
+        webSiteClient = createAzureClient(testAccount.getSubscriptionContext(), WebSiteManagementClient);
     }
-    this.timeout(2 * 60 * 1000);
-    testAccount = new TestAzureAccount(vscode);
-    await testAccount.signIn();
-    ext.azureAccountTreeItem = new AzureAccountTreeItem(testAccount);
-    ext.tree = new AzExtTreeDataProvider(ext.azureAccountTreeItem, 'appService.loadMore');
-    webSiteClient = createAzureClient(testAccount.getSubscriptionContext(), WebSiteManagementClient);
 });
 
-suiteTeardown(async function (this: IHookCallbackContext): Promise<void> {
-    if (!longRunningTestsEnabled) {
-        this.skip();
+suiteTeardown(async function (this: Mocha.Context): Promise<void> {
+    if (longRunningTestsEnabled) {
+        this.timeout(10 * 60 * 1000);
+        await Promise.all(resourceGroupsToDelete.map(async resource => {
+            await beginDeleteResourceGroup(resource);
+        }));
+        ext.azureAccountTreeItem.dispose();
     }
-    this.timeout(10 * 60 * 1000);
-    await deleteResourceGroups();
-    ext.azureAccountTreeItem.dispose();
 });
 
-async function deleteResourceGroups(): Promise<void> {
+export async function beginDeleteResourceGroup(resourceGroup: string): Promise<void> {
     const client: ResourceManagementClient = createAzureClient(testAccount.getSubscriptionContext(), ResourceManagementClient);
-    await Promise.all(resourceGroupsToDelete.map(async resourceGroup => {
-        if (await client.resourceGroups.checkExistence(resourceGroup)) {
-            console.log(`Deleting resource group "${resourceGroup}"...`);
-            await client.resourceGroups.deleteMethod(resourceGroup);
-            console.log(`Resource group "${resourceGroup}" deleted.`);
-        } else {
-            // If the test failed, the resource group might not actually exist
-            console.log(`Ignoring resource group "${resourceGroup}" because it does not exist.`);
-        }
-    }));
+    if (await client.resourceGroups.checkExistence(resourceGroup)) {
+        console.log(`Started delete of resource group "${resourceGroup}"...`);
+        await client.resourceGroups.beginDeleteMethod(resourceGroup);
+        console.log(`Successfully started delete of resource group "${resourceGroup}".`);
+    } else {
+        // If the test failed, the resource group might not actually exist
+        console.log(`Ignoring resource group "${resourceGroup}" because it does not exist.`);
+    }
 }